updating login and register with customer infomation

+++ MICROSERVICES/LTI-SPRING-MICROSERVICE-FLIPFIT-USER-PRODUCER/src/main/java/com/lti/flipfit/repository/GymAddressRepository.java
@@ -4,5 +4,5 @@ package com.lti.flipfit.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.lti.flipfit.entity.GymAddress;
 
-public interface GymAddressRepository extends JpaRepository<GymAddress, Integer> {
+public interface GymAddressRepository extends JpaRepository<GymAddress, Long> {
 }
