+++ MICROSERVICES/LTI-SPRING-MICROSERVICE-FLIPFIT-USER-PRODUCER/src/main/java/com/lti/flipfit/repository/GymRoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lti.flipfit.entity.GymRole;
 
 @Repository
-public interface GymRoleRepository extends JpaRepository<GymRole, Integer> {
+public interface GymRoleRepository extends JpaRepository<GymRole, Long> {
 
 }
